1.完成初版的开发，具备预定的所有功能，界面运行正常

+++ batterywidget.h
@@ -10,13 +10,13 @@ public:
     explicit BatteryWidget(QWidget *parent = nullptr);
     BatteryWidget();
     // 设置电池电量的接口，范围为 0 到 100
-    void setBatteryLevel(int level);
+    void setBatteryLevel(double level);
 protected:
     // 重写绘制事件，以便自定义绘制电池形状和电量
     void paintEvent(QPaintEvent *event) override;
 
 private:
-    int batteryLevel; // 当前电量，0 到 100 之间
+    double batteryLevel; // 当前电量，0 到 100 之间
 };
 
 #endif // BATTERYWIDGET_H
